add security so only store owners can change products

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -60,6 +61,7 @@ def all_products(request):
     return render(request, 'products/products.html', context)
 
 
+@login_required
 def product_detail(request, product_id):
     """ 
     A view to show individual product details 
@@ -74,10 +76,15 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ 
     A view that allows store owners to add products to the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can add products to the store.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -97,10 +104,14 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ 
     For store owners to edit a product they have added to the store 
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can edit products in the store.')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -124,10 +135,14 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ 
     Store owner can delete a product from the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can delete products from the store.')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Your Product has been deleted!')
